The first draft.
This is my first draft of the code. It uses the classes it needs for the
current code and the classes for writing to a json file and for sending
an email. its has public voids for two input fields and will load the
email paired with the name when the name is updated.